requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==9.1.1
PyYAML==6.0
scikit_learn==1.1.1
six==1.16.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
